+++ templates/header.php
@@ -1,4 +1,4 @@
-<?php // Attogram Framework - Page Header v0.1.1
+<?php // Attogram Framework - Page Header v0.1.2
 
 namespace attogram;
 
@@ -24,7 +24,7 @@ if (!isset($title) || !$title || !is_string($title)) {
 
 $navbar = $this->templates['navbar'];
 
-if (attogram_fs::is_readable_file($navbar, 'php')) {
+if (attogram_fs::isReadableFile($navbar, 'php')) {
     include $navbar;
 } else {
     $this->log->error('Missing navbar template: '.$navbar);